requirements.txt
python-telegram-bot==20.3
pytrends
gspread
oauth2client
matplotlib
